# backend/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sharobim Deacons API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/deacons/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists All Deacons Paginated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Number of elements per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "first_name, date_of_birth, country, rank_name",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ASC OR DESC",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a deacon profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "description": "Creation",
                        "name": "payload.Deacon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.Deacon"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deacons/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get A Detailed info of a specific deacon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deacon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a deacon and its skills",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "description": "Creation",
                        "name": "payload.Deacon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.Deacon"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Deacon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a Deacon and his skills",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deacon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deacons/{id}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Services done by a deacon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deacons"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deacon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetHistoryServiceByDeaconIdRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/skills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists All Skills",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.GetSkillsRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Logs In The User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "Logging",
                        "name": "payload.User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Creates a User for the application",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "Creation",
                        "name": "payload.User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "db.GetHistoryServiceByDeaconIdRow": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "liturgyName": {
                    "type": "string"
                },
                "skill": {
                    "type": "string"
                }
            }
        },
        "db.GetSkillsRow": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "skill": {
                    "type": "string"
                }
            }
        },
        "payload.Deacon": {
            "type": "object",
            "required": [
                "country",
                "deacon_rank",
                "first_name",
                "last_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "deacon_rank": {
                    "type": "integer"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/payload.DeaconSkill"
                    }
                }
            }
        },
        "payload.DeaconSkill": {
            "type": "object",
            "required": [
                "score",
                "skill_id"
            ],
            "properties": {
                "score": {
                    "type": "integer"
                },
                "skill_id": {
                    "type": "integer"
                }
            }
        },
        "payload.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "response.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}